Fixed client issues when logging failing
1. When you call LogSync method and Loggly server returns an error (i.e.
http status code 500), the calling code will be hanging because callback
function will be never called (see Logger class changes). The main idea
was to add the logging status to the LogResponse class and to assume
that callback must be called in any case.
2. When you call LogSync method and Loggly server is unavailable due to
some network issues, the callback function won't be called as well (see
above). See Communicator class fix for that