support ForwardedForIp property in transport object
There is a scenario where the loggly client is being used on desktop applications, and therefore the client's IP is being sent to loggly and indexed as clientHost. In some cases, this IP can be considered PII and undesired. By creating a new property on the transport object (ForwardedForIp), the requestes made to loggly can contain a new IP (or value) to be sent and override the client's IP.